Creating + using utils.py to house functions used by both workflows.
